requirements.txt
streamlit
tensorflow
opencv-python-headless
matplotlib
pillow
numpy
gdown